% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libr.R
\name{display.libr.all}
\alias{display.libr.all}
\title{Display all LIBR palettes in a stacked plot}
\usage{
display.libr.all()
}
\value{
A \pkg{patchwork} plot showing every palette.
}
\description{
Display all LIBR palettes in a stacked plot
}
\examples{
display.libr.all()
}
