% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libr.R
\name{libr.pal}
\alias{libr.pal}
\title{Retrieve a LIBR color palette}
\usage{
libr.pal(n, name)
}
\arguments{
\item{n}{Integer; number of colors to return.}

\item{name}{Character; one of the named palettes in \code{libr_themes$palette}. Continuous gradient options include: \code{"LIBR_logo_sunset"}, \code{"LIBR_logo_sky"}, \code{"LIBR_logo_purple"}, \code{"LIBR_logo_pink"}, \code{"LIBR_gradient_purple"}, \code{"LIBR_gradient_orange"}, \code{"LIBR_gradient_pink"}, \code{"LIBR_gradient_yellow"}, and \code{"LIBR_gradient_blue"}. Discrete palette options include: \code{"LIBR_theme_dark"}, \code{"LIBR_theme_bright"}, \code{"LIBR_theme_pastel"}, \code{"LIBR_theme_dark_expanded"}, \code{"LIBR_theme_bright_expanded"}, and \code{"LIBR_theme_pastel_expanded"}.}
}
\value{
A character vector of HEX color codes.
}
\description{
Retrieve a LIBR color palette
}
\examples{
libr.pal(3, "LIBR_theme_dark")
}
